Hyeonss/post

+++ src/components/pages/PostDetail/PostDetailContainer.tsx
@@ -59,7 +59,9 @@ const PostDetailContainer = () => {
     history.push("/write");
   }, [dispatch, history, postData]);
   const handleDeletePost = useCallback(async () => {
-    console.log(postData?.getPostDetail);
+    const isConfirmed = confirm("정말로 게시물을 삭제하시겠습니까?");
+    if (!isConfirmed) return false;
+
     const data = postData?.getPostDetail;
     await deletePostMutation({
       variables: {
